@@ -21,13 +21,13 @@
   "classmap-authoritative": true,
   "prefer-stable": true,
   "require": {
-    "php": ">=5.6.0",
-    "zendframework/zend-session": "2.8.*",
-    "zendframework/zend-cache": "2.8.*",
-    "zendframework/zend-serializer": "2.9.*",
-    "zendframework/zend-validator": "2.10.*",
-    "zendframework/zend-mail": "2.10.*",
-    "zendframework/zend-i18n": "2.9.*",
+    "php": ">=7.0",
+    "laminas/laminas-session": "2.9.*",
+    "laminas/laminas-cache": "2.9.*",
+    "laminas/laminas-serializer": "2.9.*",
+    "laminas/laminas-validator": "2.12.*",
+    "laminas/laminas-mail": "2.10.*",
+    "laminas/laminas-i18n": "2.10.*",
     "zf1/zend-registry": "~1.12",
     "zf1/zend-config": "~1.12",
     "zf1/zend-acl": "~1.12",
@@ -38,10 +38,9 @@
     "belhard/floatthead": "dev-master",
     "belhard/simplemodal": "dev-master",
     "components/jquery": "2.1.4",
-    "monolog/monolog": "1.23.*",
-    "guzzlehttp/guzzle": "6.5.0",
+    "monolog/monolog": "1.25.*",
+    "guzzlehttp/guzzle": "6.5.*",
     "blueimp/jquery-file-upload": "10.*",
-    "firephp/firephp-core": "*",
     "phpseclib/phpseclib": "~2.0",
     "ext-json": "*",
     "ext-zip": "*",
