@@ -2,8 +2,8 @@
 namespace Core2;
 
 require_once "Cache.php";
-use Zend\Cache\StorageFactory;
-
+use Laminas\Cache\StorageFactory;
+use Laminas\Session\Container as SessionContainer;
 
 /**
  * Class Db
@@ -162,17 +162,6 @@ class Db {
 
 
     /**
-     * Ищет перевод для строки $str
-     *
-     * @param string $str
-     * @return string
-     */
-    public function _($str) {
-        return $this->translate->tr($str);
-	}
-
-
-    /**
      * @param \Zend_Config $database
      * @return \Zend_Db_Adapter_Abstract
      */
@@ -309,7 +298,7 @@ class Db {
      */
 	public function closeSession($expired = 'N') {
 
-		$auth = new \Zend\Session\Container('Auth');
+		$auth = new SessionContainer('Auth');
 
 		if ($auth && $auth->ID && $auth->LIVEID) {
             $row = $this->dataSession->find($auth->LIVEID)->current();
@@ -332,7 +321,7 @@ class Db {
 	 */
 	public function logActivity($exclude = array()) {
 
-        $auth = new \Zend\Session\Container('Auth');
+        $auth = new SessionContainer('Auth');
 
         if ($auth->ID && $auth->ID > 0 && $auth->LIVEID) {
             if ($exclude && in_array($_SERVER['QUERY_STRING'], $exclude)) {
@@ -363,7 +352,7 @@ class Db {
                 $this->config->log->system->writer == 'file'
             ) {
                 if ( ! $this->config->log->system->file) {
-                    throw new \Exception($this->_('Не задан файл журнала запросов'));
+                    throw new \Exception($this->translate->tr('Не задан файл журнала запросов'));
                 }
 
                 $log = new Log('access');
@@ -502,6 +491,17 @@ class Db {
 		}
 		return $res;
 	}
+
+
+    /**
+     * Ищет перевод для строки $str
+     * @param string $str
+     * @param string $module
+     * @return string
+     */
+    public function _($str, $module = 'core2') {
+        return $this->translate->tr($str, $module);
+    }
 
 
 	/**
