@@ -8,7 +8,7 @@
  */
 namespace Core2\Store;
 
-use Zend\Session\Container as SessionContainer;
+use Laminas\Session\Container as SessionContainer;
 
 require_once(DOC_ROOT . "core2/inc/classes/Common.php");
 require_once(DOC_ROOT . "core2/inc/classes/Image.php");
