@@ -7,7 +7,7 @@ use Monolog\Handler\SlackWebhookHandler;
 use Monolog\Handler\SyslogHandler;
 use Monolog\Processor\WebProcessor;
 use Monolog\Formatter\NormalizerFormatter;
-use Zend\Session\Container as SessionContainer;
+use Laminas\Session\Container as SessionContainer;
 
 
 /**
