@@ -1,7 +1,7 @@
 <?php
 namespace Core2;
 
-use Zend\Session\Container as SessionContainer;
+use Laminas\Session\Container as SessionContainer;
 
 
 require_once 'Templater3.php';
@@ -29,100 +29,125 @@ class Login extends Db {
      */
     public function dispatch() {
 
-        if (isset($_GET['core']) && $this->config->mail && $this->config->mail->server) {
-            if ($this->core_config->registration &&
-                $this->core_config->registration->on &&
-                $this->core_config->registration->role_id
-            ) {
+        if (isset($_GET['core'])) {
+            if ($this->config->mail && $this->config->mail->server) {
+                if ($this->core_config->registration &&
+                    $this->core_config->registration->on &&
+                    $this->core_config->registration->role_id
+                ) {
 
-                if ($_GET['core'] == 'registration') {
-                    if ($_SERVER['REQUEST_METHOD'] === 'GET') {
-                        return $this->getRegistration();
+                    if ($_GET['core'] == 'registration') {
+                        if ($_SERVER['REQUEST_METHOD'] === 'GET') {
+                            return $this->getPageRegistration();
 
-                    } elseif ($_SERVER['REQUEST_METHOD'] === 'POST') {
-                        $data = [
-                            'company_name' => isset($_POST['company_name']) ? $_POST['company_name'] : '',
-                            'email'        => isset($_POST['email']) ? $_POST['email'] : '',
-                            'unp'          => isset($_POST['unp']) ? $_POST['unp'] : '',
-                            'tel'          => isset($_POST['tel']) ? $_POST['tel'] : '',
-                        ];
+                        } elseif ($_SERVER['REQUEST_METHOD'] === 'POST') {
+                            return $this->registration($_POST);
 
-                        return $this->registration($data);
+                        } else {
+                            http_response_code(404);
 
-                    } else {
-                        http_response_code(404);
-                        return '';
+                            return '';
+                        }
+                    }
+
+                    if ($_GET['core'] == 'registration_complete') {
+                        if ($_SERVER['REQUEST_METHOD'] === 'GET') {
+                            if (empty($_GET['key'])) {
+                                http_response_code(404);
+                                return '';
+                            }
+
+                            return $this->getPageRegistrationComplete($_GET['key']);
+
+                        } elseif ($_SERVER['REQUEST_METHOD'] === 'POST') {
+                            if (empty($_POST['key'])) {
+                                http_response_code(404);
+                                return '';
+                            }
+                            if (empty($_POST['password'])) {
+                                return json_encode([
+                                    'status'        => 'error',
+                                    'error_message' => $this->_('Заполните пароль')
+                                ]);
+                            }
+
+                            return $this->registrationComplete($_POST['key'], $_POST['password']);
+
+                        } else {
+                            http_response_code(404);
+                            return '';
+                        }
                     }
                 }
 
-                if ($_GET['core'] == 'registration_complete') {
-                    if ($_SERVER['REQUEST_METHOD'] === 'GET') {
-                        if (empty($_GET['key'])){
+                if ($this->core_config->restore && $this->core_config->restore->on) {
+                    if ($_GET['core'] == 'restore') {
+                        if ($_SERVER['REQUEST_METHOD'] === 'GET') {
+                            return $this->getPageRestore();
+
+                        } elseif ($_SERVER['REQUEST_METHOD'] === 'POST') {
+                            if (empty($_POST['email'])) {
+                                return json_encode([
+                                    'status'        => 'error',
+                                    'error_message' => $this->_('Заполните email')
+                                ]);
+                            }
+
+                            return $this->restore($_POST["email"]);
+
+                        } else {
                             http_response_code(404);
                             return '';
                         }
-                        return $this->registrationComplete($_GET['key']);
+                    }
 
-                    } elseif ($_SERVER['REQUEST_METHOD'] === 'POST') {
-                        if (empty($_POST['key'])){
-                            http_response_code(404);
-                            return '';
+
+                    if ($_GET['core'] == 'restore_complete') {
+                        if ($_SERVER['REQUEST_METHOD'] === 'GET') {
+                            if (empty($_GET['key'])) {
+                                http_response_code(404);
+                                return '';
+                            }
+
+                            return $this->getPageRestoreComplete($_GET['key']);
+
+                        } elseif ($_SERVER['REQUEST_METHOD'] === 'POST') {
+                            if (empty($_POST['key'])) {
+                                http_response_code(404);
+                                return '';
+                            }
+
+                            if (empty($_POST['password'])) {
+                                return json_encode([
+                                    'status'        => 'error',
+                                    'error_message' => $this->_('Заполните пароль')
+                                ]);
+                            }
+
+                            return $this->restoreComplete($_POST['key'], $_POST['password']);
                         }
-                        if (empty($_POST['password'])) {
-                            return json_encode([
-                                'status'  => 'error',
-                                'message' => $this->_('Заполните пароль')
-                            ]);
-                        }
-
-                        return $this->registrationPass($_POST['key'], $_POST['password']);
-
-                    } else {
-                        http_response_code(404);
-                        return '';
                     }
                 }
             }
 
-            if ($this->core_config->restore && $this->core_config->restore->on) {
-                if ($_GET['core'] == 'restore') {
-                    if ($_SERVER['REQUEST_METHOD'] === 'GET') {
 
-                        if ( ! empty($_GET['key'])) {
-                            return $this->restoreComplete($_GET['key']);
-
-                        } else {
-                            return $this->getRestore();
-                        }
-
-                    } elseif ($_SERVER['REQUEST_METHOD'] === 'POST') {
-                        if (empty($_POST['email'])) {
-                            return json_encode([
-                                'status'  => 'error',
-                                'message' => $this->_('Заполните email')
-                            ]);
-                        }
-
-                        return $this->restore($_POST["email"]);
-
-                    } else {
-                        http_response_code(404);
-                        return '';
+            if ($_GET['core'] == 'login') {
+                if ($_SERVER['REQUEST_METHOD'] === 'POST') {
+                    if (empty($_POST['login'])) {
+                        return json_encode([
+                            'status'  => 'error',
+                            'error_message' => $this->_('Заполните логин')
+                        ]);
                     }
-                }
 
-                if ($_GET['core'] == 'restore_complete') {
-                    if (empty($_POST['key'])){
-                        http_response_code(404);
-                        return '';
-                    }
                     if (empty($_POST['password'])) {
                         return json_encode([
                             'status'  => 'error',
-                            'message' => $this->_('Заполните пароль')
+                            'error_message' => $this->_('Заполните пароль')
                         ]);
                     }
-                    return $this->restorePass($_POST['key'], $_POST['password']);
+
+                    return $this->login($_POST["login"], $_POST['password']);
                 }
             }
         }
@@ -138,7 +163,7 @@ class Login extends Db {
             }
         }
 
-        return $this->getLogin();
+        return $this->getPageLogin();
     }
 
 
@@ -160,10 +185,10 @@ class Login extends Db {
      * @param $favicon
      * @throws \Exception
      */
-    public function setFavicon($favicon) {
+    public function setFavicon(Array $favicon) {
 
         if ( ! is_array($favicon)) {
-            throw new \Exception('Incorrect favicon data');
+            throw new \Exception('Incorrect favicon');
         }
 
         $this->favicon = $favicon;
@@ -176,54 +201,14 @@ class Login extends Db {
      * @throws \Zend_Exception
      * @throws \Exception
      */
-    private function getLogin() {
-
-        if (isset($_POST['action'])) {
-            require_once 'core2/inc/CoreController.php';
-            $this->setContext('admin');
-            $core = new \CoreController();
-
-            $url  = "index.php";
-            if ($core->action_login($_POST) && ! empty($_SERVER['QUERY_STRING'])) {
-                $url .= "#" . $_SERVER['QUERY_STRING'];
-            }
-
-            header("Location: $url");
-            return '';
-        }
+    private function getPageLogin() {
 
         $tpl = new \Templater2("core2/html/" . THEME . "/login/login.html");
-
-        $errorNamespace = new SessionContainer('Error');
-        $blockNamespace = new SessionContainer('Block');
-
-        if ( ! empty($blockNamespace->blocked)) {
-            $tpl->error->assign('[ERROR_MSG]', $errorNamespace->ERROR);
-            $tpl->assign('[ERROR_LOGIN]', '');
-        } elseif ( ! empty($errorNamespace->ERROR)) {
-            $tpl->error->assign('[ERROR_MSG]', $errorNamespace->ERROR);
-            $tpl->assign('[ERROR_LOGIN]', $errorNamespace->TMPLOGIN);
-            $errorNamespace->ERROR = '';
-        } else {
-            $tpl->error->assign('[ERROR_MSG]', '');
-            $tpl->assign('[ERROR_LOGIN]', '');
-        }
-
-        if (empty($this->config->ldap->active) || ! $this->config->ldap->active) {
-            $tpl->assign('<form', "<form onsubmit=\"document.getElementById('gfhjkm').value=hex_md5(document.getElementById('gfhjkm').value)\"");
-        }
 
         $logo = $this->getSystemLogo();
 
         if (is_file($logo)) {
             $tpl->logo->assign('{logo}', $logo);
-        }
-
-        $reg  = \Zend_Registry::getInstance();
-        $auth = $reg->get('auth');
-
-        if ( ! empty($auth->TOKEN)) {
-            $tpl->assign('name="action"', 'name="action" value="' . $auth->TOKEN . '"');
         }
 
         if ($this->config->mail && $this->config->mail->server) {
@@ -252,7 +237,7 @@ class Login extends Db {
      * @return string
      * @throws \Exception
      */
-    private function getRegistration() {
+    private function getPageRegistration() {
 
         $tpl  = new \Templater3("core2/html/" . THEME . "/login/registration.html");
         $logo = $this->getSystemLogo();
@@ -267,154 +252,43 @@ class Login extends Db {
             }
         }
 
+        $isset_phone = false;
+
+        if ($this->core_config->registration->fields) {
+            $fields = $this->core_config->registration->fields->toArray();
+
+            if ( ! empty($fields)) {
+                foreach ($fields as $name => $field) {
+
+                    $type = ! empty($field['type']) ? htmlspecialchars($field['type']) : 'text';
+
+                    if ($type == 'phone') {
+                        $isset_phone = true;
+                    }
+
+                    $tpl->field->assign('[NAME]',     $name);
+                    $tpl->field->assign('[TYPE]',     $type);
+                    $tpl->field->assign('[TITLE]',    ! empty($field['title']) ? htmlspecialchars($field['title']) : '');
+                    $tpl->field->assign('[REQUIRED]', ! empty($field['required']) ? 'required' : '');
+                    $tpl->field->reassign();
+                }
+            }
+        }
+
 
         $html = $this->getIndex();
         $html = str_replace('<!--index -->', $tpl->render(), $html);
 
+
+        if ($isset_phone) {
+            $scripts = [
+                '<script src="core2/js/cleave.min.js"></script>',
+                '<script src="core2/js/cleave-phone.i18n.js"></script>',
+            ];
+            $html = str_replace('<!--system_js-->', implode('', $scripts), $html);
+        }
+
         return $html;
-    }
-
-
-    /**
-     * @param $data
-     * @return false|string
-     * @throws \Zend_Db_Adapter_Exception
-     * @throws \Zend_Db_Exception
-     */
-    private function registration(array $data) {
-
-        if (empty($data['email'])) {
-            return json_encode([
-                'status'  => 'error',
-                'message' => $this->_('Email не заполнен')
-            ]);
-        }
-
-        if (empty($data['unp'])) {
-            return json_encode([
-                'status'  => 'error',
-                'message' => $this->_('УНП не заполнено')
-            ]);
-        }
-
-        if (empty($data['company_name'])) {
-            return json_encode([
-                'status'  => 'error',
-                'message' => $this->_('Название организации не заполнено')
-            ]);
-        }
-
-        if (empty($data['tel'])) {
-            return json_encode([
-                'status'  => 'error',
-                'message' => $this->_('Телефон не заполнен')
-            ]);
-        }
-
-        $data['email'] = trim($data['email']);
-
-        $isset_user_login = $this->db->fetchOne("
-            SELECT 1
-            FROM core_users
-            WHERE u_login = ?
-        ", $data['email']);
-
-
-        if ($isset_user_login) {
-            return json_encode([
-                'status'  => 'error',
-                'message' => $this->_('Такой пользователь уже есть')
-            ]);
-        }
-
-        $isset_user_email = $this->db->fetchOne("
-            SELECT 1
-            FROM core_users
-            WHERE email = ?
-        ", $data['email']);
-
-        if ($isset_user_email) {
-            return json_encode([
-                'status'  => 'error',
-                'message' => $this->_('Пользователь с таким Email уже есть')
-            ]);
-        }
-
-        $isset_unp = $this->db->fetchOne("
-            SELECT 1
-            FROM mod_ordering_contractors
-            WHERE unp = ?
-              AND is_deleted_sw = 'N' 
-              AND user_id IS NOT NULL
-        ", $data['unp']);
-
-        if ($isset_unp) {
-            return json_encode([
-                'status'  => 'error',
-                'message' => $this->_('Организация с таким УНП уже есть')
-            ]);
-        }
-
-        $isset_email = $this->db->fetchOne("
-            SELECT 1
-            FROM mod_ordering_contractors
-            WHERE email = ?
-              AND is_deleted_sw = 'N' 
-              AND user_id IS NOT NULL
-        ", $data['email']);
-
-        if ($isset_email) {
-            return json_encode([
-                'status'  => 'error',
-                'message' => $this->_('Организация с таким email уже есть')
-            ]);
-        }
-
-
-        $contractor = $this->db->fetchRow("
-            SELECT id,
-                   email,
-                   is_active_sw
-            FROM mod_ordering_contractors
-            WHERE (email = ? OR unp = ?)
-              AND user_id IS NULL
-        ", [
-            $data['email'],
-            $data['unp']
-        ]);
-
-        if ( ! empty($contractor)) {
-            $reg_key = \Tool::pass_salt(md5($data['email'] . microtime()));
-            $where   = $this->db->quoteInto('id = ?', $contractor['id']);
-            $this->db->update('mod_ordering_contractors', [
-                'reg_key'      => $reg_key,
-                'date_expired' => new \Zend_Db_Expr('DATE_ADD(NOW(), INTERVAL 1 DAY)')
-            ], $where);
-
-            $this->sendEmailRegistration($data['email'], $reg_key);
-
-            return json_encode([
-                'status'  => 'success',
-                'message' => $this->_('На указанную вами почту отправлены данные для входа в систему')
-            ]);
-        }
-
-        $reg_key = \Tool::pass_salt(md5($data['email'] . microtime()));
-        $this->db->insert('mod_ordering_contractors', [
-            'title'        => $data['company_name'],
-            'email'        => $data['email'],
-            'unp'          => $data['unp'],
-            'phone'        => $data['tel'],
-            'reg_key'      => $reg_key,
-            'date_expired' => new \Zend_Db_Expr('DATE_ADD(NOW(), INTERVAL 1 DAY)'),
-        ]);
-
-        $this->sendEmailRegistration($data['email'], $reg_key);
-
-        return json_encode([
-            'status'  => 'success',
-            'message' => $this->_('На указанную вами почту отправлены данные для входа в систему')
-        ]);
     }
 
 
@@ -424,7 +298,7 @@ class Login extends Db {
      * @throws \Zend_Db_Exception
      * @throws \Exception
      */
-    private function registrationComplete($key) {
+    private function getPageRegistrationComplete($key) {
 
         $tpl  = new \Templater3("core2/html/" . THEME . "/login/registration-complete.html");
         $logo = $this->getSystemLogo();
@@ -435,9 +309,10 @@ class Login extends Db {
 
         $isset_key = $this->db->fetchOne("
             SELECT 1
-            FROM mod_ordering_contractors
+            FROM core_users
             WHERE reg_key = ?
               AND date_expired > NOW()
+              AND visible = 'N'
         ", $key);
 
         $error_message = '';
@@ -464,108 +339,10 @@ class Login extends Db {
 
 
     /**
-     * @param $key
-     * @param $password
-     * @return false|string
-     * @throws \Zend_Db_Adapter_Exception
-     * @throws \Zend_Db_Exception
-     */
-    private function registrationPass($key, $password) {
-
-        $user_info = $this->db->fetchRow("
-            SELECT id,
-                   user_id,
-                   email,
-                   title
-            FROM mod_ordering_contractors 
-            WHERE reg_key = ?
-              AND date_expired > NOW()
-            LIMIT 1
-        ", $key);
-
-        if (empty($user_info)) {
-            return json_encode([
-                'status'  => 'error',
-                'message' => $this->_('Ссылка устарела')
-            ]);
-        }
-
-        $isset_login = $this->db->fetchOne("
-            SELECT 1
-            FROM core_users AS u 
-            WHERE u_login = ?
-        ", trim($user_info['email']));
-
-        if ($isset_login) {
-            return json_encode([
-                'status'  => 'error',
-                'message' => $this->_('Пользователь с таким логином уже существует')
-            ]);
-        }
-
-        $isset_email = $this->db->fetchOne("
-            SELECT 1
-            FROM core_users AS u 
-            WHERE u_login = ?
-        ", trim($user_info['email']));
-
-        if ($isset_email) {
-            return json_encode([
-                'status'  => 'error',
-                'message' => $this->_('Пользователь с таким email уже существует')
-            ]);
-        }
-
-        $this->db->beginTransaction();
-        try {
-            $this->db->insert('core_users', [
-                'visible'     => 'Y',
-                'is_admin_sw' => 'N',
-                'u_login'     => trim($user_info['email']),
-                'u_pass'      => \Tool::pass_salt($password),
-                'email'       => trim($user_info['email']),
-                'date_added'  => new \Zend_Db_Expr('NOW()'),
-                'role_id'     => $this->core_config->registration->role_id
-            ]);
-            $user_id = $this->db->lastInsertId();
-
-            $this->db->insert('core_users_profile', [
-                'user_id'   => $user_id,
-                'firstname' => $user_info['title'],
-            ]);
-
-            $where = $this->db->quoteInto('id = ?', $user_info['id']);
-            $this->db->update('mod_ordering_contractors', [
-                'reg_key'      => new \Zend_Db_Expr('NULL'),
-                'date_expired' => new \Zend_Db_Expr('NULL'),
-                'user_id'      => $user_id,
-                'is_active_sw' => 'Y',
-            ], $where);
-
-            $this->db->commit();
-
-        } catch (\Exception $e) {
-            $this->db->rollback();
-
-            return json_encode([
-                'status'  => 'error',
-                'message' => "Ошибка создания пользователя. Обратитесь к администратору"
-            ]);
-        }
-
-
-        return json_encode([
-            'status'  => 'success',
-            'message' => '<h4>Готово!</h4><p>Вы можете войти в систему</p>'
-        ]);
-    }
-
-
-    /**
      * @return string|string[]
      * @throws \Exception
      */
-    private function getRestore() {
+    private function getPageRestore() {
 
         $tpl = new \Templater3("core2/html/" . THEME . "/login/restore.html");
 
@@ -592,71 +369,11 @@ class Login extends Db {
 
 
     /**
-     * @param $email
-     * @return false|string|string[]
-     * @throws \Zend_Db_Adapter_Exception
-     * @throws \Zend_Db_Exception
-     */
-    private function restore($email) {
-
-        $user_id = $this->db->fetchOne("
-            SELECT u.u_id
-            FROM core_users AS u
-                JOIN mod_ordering_contractors AS oc ON u.u_id = oc.user_id
-            WHERE u.email = ?
-            LIMIT 1
-        ", $email);
-
-        if (empty($user_id)) {
-            return json_encode([
-                'status'  => 'error',
-                'message' => $this->_('Пользователя с таким Email нет в системе')
-            ]);
-        }
-
-
-        $reg_key = \Tool::pass_salt(md5($email . microtime()));
-        $where   = $this->db->quoteInto('user_id = ?', $user_id);
-        $this->db->update('mod_ordering_contractors', [
-            'reg_key'      => $reg_key,
-            'date_expired' => new \Zend_Db_Expr('DATE_ADD(NOW(), INTERVAL 1 DAY)')
-        ], $where);
-
-
-        $protocol = ! empty($this->config->system) && $this->config->system->https ? 'https' : 'http';
-        $host     = ! empty($this->config->system) ? $this->config->system->host : '';
-        $doc_path = rtrim(DOC_PATH, '/') . '/';
-
-        $content_email = "
-            Вы запросили смену пароля на сервисе {$host}<br>
-            Для продолжения <b>перейдите по указанной ниже ссылке</b>.<br><br>
-
-            <a href=\"{$protocol}://{$host}{$doc_path}index.php?core=restore&key={$reg_key}\" 
-               style=\"font-size: 16px\">{$protocol}://{$host}{$doc_path}index.php?core=restore&key={$reg_key}</a>
-        ";
-
-        $this->setContext('queue');
-
-        require_once 'Email.php';
-        $core_email = new \Core2\Email();
-        $core_email->to($email)
-            ->subject("Автопромсервис: Восстановление пароля")
-            ->body($content_email)
-            ->send(true);
-
-        return json_encode([
-            'status'  => 'success',
-            'message' => $this->_('На указанную вами почту отправлены данные для смены пароля')
-        ]);
-    }
-
-
-    /**
      * @param $key
      * @return string|string[]
      * @throws \Exception
      */
-    private function restoreComplete($key) {
+    private function getPageRestoreComplete($key) {
 
         $tpl = new \Templater3("core2/html/" . THEME . "/login/restore-complete.html");
 
@@ -668,7 +385,7 @@ class Login extends Db {
 
         $isset_key = $this->db->fetchOne("
             SELECT 1
-            FROM mod_ordering_contractors
+            FROM core_users
             WHERE reg_key = ?
               AND date_expired > NOW()
         ", $key);
@@ -701,17 +418,378 @@ class Login extends Db {
 
 
     /**
+     * Авторизация пользователя через форму
+     * @param $login
+     * @param $password
+     * @return string
+     */
+    private function login($login, $password) {
+
+        $blockNamespace = new SessionContainer('Block');
+
+        try {
+
+            if ( ! empty($blockNamespace->blocked)) {
+                throw new \Exception($this->translate->tr("Ваш доступ временно заблокирован!"));
+            }
+
+
+            $login = trim($login);
+
+
+            if ($this->config->ldap &&
+                $this->config->ldap->active &&
+                ((function_exists('ctype_print') ? ! ctype_print($password) : true) || strlen($password) < 30)
+            ) {
+                throw new \Exception($this->translate->tr("Ошибка пароля!"), 400);
+            }
+
+
+            $this->getConnection($this->config->database);
+
+
+            $auth         = [];
+            $is_ldap_auth = false;
+
+            if ($login === 'root') {
+                $auth = $this->getAuthRoot();
+
+            } else {
+                // LDAP
+                if ($this->config->ldap && $this->config->ldap->active) {
+                    $auth = $this->getAuthLdap($login, $password);
+
+                    if ( ! empty($auth)) {
+                        $is_ldap_auth = true;
+
+                    } else {
+                        $password = md5($password);
+                    }
+                }
+
+                // LOGIN
+                if (empty($auth)) {
+                    $auth = $this->getAuthLogin($login);
+                }
+            }
+
+            if ( ! $auth) {
+                throw new \Exception($this->translate->tr("Нет такого пользователя"));
+            }
+
+
+            $auth['LDAP'] = $is_ldap_auth;
+
+            $md5_pass = \Tool::pass_salt($password);
+
+
+            if ($auth['LDAP']) {
+                $auth['u_pass'] = $md5_pass;
+            }
+
+            if ($auth['u_pass'] !== $md5_pass) {
+                throw new \Exception($this->translate->tr("Неверный пароль"));
+            }
+
+            $authNamespace = new SessionContainer('Auth');
+            $authNamespace->accept_answer = true;
+
+            $session_life = $this->db->fetchOne("
+                SELECT value 
+                FROM core_settings 
+                WHERE visible = 'Y' 
+                  AND code = 'session_lifetime' 
+                LIMIT 1
+            ");
+
+            if ($session_life) {
+                $authNamespace->setExpirationSeconds($session_life, "accept_answer");
+            }
+
+            if (session_id() == 'deleted') {
+                throw new \Exception($this->translate->tr("Ошибка сохранения сессии. Проверьте настройки системного времени."));
+            }
+
+            $authNamespace->ID    = (int)$auth['u_id'];
+            $authNamespace->NAME  = $auth['u_login'];
+            $authNamespace->EMAIL = $auth['email'];
+
+            if ($auth['u_login'] == 'root') {
+                $authNamespace->ADMIN  = true;
+                $authNamespace->ROLEID = 0;
+            } else {
+                $authNamespace->LN     = $auth['lastname'];
+                $authNamespace->FN     = $auth['firstname'];
+                $authNamespace->MN     = $auth['middlename'];
+                $authNamespace->ADMIN  = $auth['is_admin_sw'] == 'Y';
+                $authNamespace->ROLE   = $auth['role'] ? $auth['role'] : -1;
+                $authNamespace->ROLEID = $auth['role_id'] ? $auth['role_id'] : 0;
+                $authNamespace->LIVEID = $this->storeSession($authNamespace);
+            }
+
+            $authNamespace->LDAP = $auth['LDAP'];
+
+
+            //регенерация сессии для предотвращения угона
+            if ( ! ($authNamespace->init)) {
+                $authNamespace->getManager()->regenerateId();
+                $authNamespace->init = true;
+            }
+
+
+            return json_encode([
+                'status' => 'success'
+            ]);
+
+        } catch (\Exception $e) {
+            $code = $e->getCode() > 200 && $e->getCode() < 600 ? $e->getCode() : 403;
+            http_response_code($code);
+
+            if (isset($blockNamespace->numberOfPageRequests)) {
+                $blockNamespace->numberOfPageRequests++;
+            } else {
+                $blockNamespace->numberOfPageRequests = 1;
+            }
+
+            if ($blockNamespace->numberOfPageRequests > 5) {
+                $blockNamespace->blocked = time();
+                $blockNamespace->setExpirationSeconds(60);
+                $blockNamespace->numberOfPageRequests = 1;
+            }
+
+
+            return json_encode([
+                'status'        => 'error',
+                'error_message' => $e->getMessage(),
+            ]);
+        }
+    }
+
+
+    /**
+     * @param $data
+     * @return false|string
+     * @throws \Zend_Db_Adapter_Exception
+     * @throws \Zend_Db_Exception
+     * @throws \Exception
+     */
+    private function registration(array $data) {
+
+        // Кастомная регистрация
+        if ($this->core_config->registration->module) {
+            $module_name = strtolower($this->core_config->registration->module);
+            $location    = $this->getModuleLocation($module_name);
+
+            $mod_controller_name = "Mod" . ucfirst($module_name) . "Controller";
+
+            if ( ! file_exists("{$location}/{$mod_controller_name}.php")) {
+                throw new \Exception(sprintf($this->_('Контроллер модуля %s не найден'), $module_name));
+            }
+
+            require_once "{$location}/{$mod_controller_name}.php";
+
+
+            $this->setContext($module_name);
+            $mod_controller = new $mod_controller_name();
+            if ( ! ($mod_controller instanceof \Registration)) {
+                throw new \Exception(sprintf($this->_('Контроллер модуля %s не поддерживает регистрацию'), $module_name));
+            }
+
+            try {
+                $result_text = $mod_controller->coreRegistration($data);
+
+                if ($result_text && is_string($result_text)) {
+                    return json_encode([
+                        'status'  => 'success',
+                        'message' => $result_text,
+                    ]);
+
+                } else {
+                    throw new \Exception($this->_('Не удалось завершить регистрацию. Попробуйте позже, либо свяжитесь с администратором'));
+                }
+
+            } catch (\Exception $e) {
+                return json_encode([
+                    'status'        => 'error',
+                    'error_message' => $e->getMessage(),
+                ]);
+            }
+        }
+
+
+        // Стандартная регистрация
+        if (empty($data['name'])) {
+            return json_encode([
+                'status'        => 'error',
+                'error_message' => $this->_('Имя не заполнено'),
+            ]);
+        }
+
+        if (empty($data['login'])) {
+            return json_encode([
+                'status'        => 'error',
+                'error_message' => $this->_('Логин не заполнен'),
+            ]);
+        }
+
+        if (empty($data['email'])) {
+            return json_encode([
+                'status'        => 'error',
+                'error_message' => $this->_('Email не заполнен'),
+            ]);
+        }
+
+        $data['name']  = trim($data['name']);
+        $data['email'] = trim($data['email']);
+        $data['login'] = trim($data['login']);
+
+
+        $isset_user_login = $this->db->fetchOne("
+            SELECT 1
+            FROM core_users
+            WHERE u_login = ?
+              AND visible = 'Y'
+        ", $data['login']);
+
+
+        if ($isset_user_login) {
+            return json_encode([
+                'status'        => 'error',
+                'error_message' => $this->_('Пользователь с таким логином уже есть'),
+            ]);
+        }
+
+        $isset_user_email = $this->db->fetchOne("
+            SELECT 1
+            FROM core_users
+            WHERE email = ?
+              AND visible = 'Y'
+        ", $data['email']);
+
+        if ($isset_user_email) {
+            return json_encode([
+                'status'        => 'error',
+                'error_message' => $this->_('Пользователь с таким email уже есть'),
+            ]);
+        }
+
+
+        $user = $this->db->fetchRow("
+            SELECT u_id AS id,
+                   email,
+                   visible AS is_active_sw
+            FROM core_users
+            WHERE (email = ? OR u_login = ?)
+              AND visible = 'N'
+        ", [
+            $data['email'],
+            $data['login']
+        ]);
+
+        $reg_key = md5($data['email'] . microtime());
+
+        if ( ! empty($user)) {
+            $where   = $this->db->quoteInto('u_id = ?', $user['id']);
+            $this->db->update('core_users', [
+                'reg_key'      => $reg_key,
+                'date_expired' => new \Zend_Db_Expr('DATE_ADD(NOW(), INTERVAL 1 DAY)')
+            ], $where);
+
+        } else {
+            $this->db->insert('core_users', [
+                'u_login'        => $data['login'],
+                'email'          => $data['email'],
+                'role_id'        => $this->core_config->registration->role_id,
+                'visible'        => 'N',
+                'is_email_wrong' => 'N',
+                'reg_key'        => $reg_key,
+                'date_expired'   => new \Zend_Db_Expr('DATE_ADD(NOW(), INTERVAL 1 DAY)'),
+                'date_added'     => new \Zend_Db_Expr('NOW()'),
+            ]);
+
+            $user_id = $this->db->lastInsertId();
+
+            $data['name'] = preg_replace('~[ ]{2,}~', ' ', $data['name']);
+
+            $name_explode = explode(' ', $data['name']);
+            $middlename   = ! empty($name_explode[2]) ? $name_explode[2] : '';
+            $lastname     = ! empty($name_explode[1]) ? $name_explode[0] : '';
+            $firstname    = $lastname ? $name_explode[1] : $name_explode[0];
+
+            $this->db->insert('core_users_profile', [
+                'user_id'    => $user_id,
+                'lastname'   => $lastname,
+                'firstname'  => $firstname,
+                'middlename' => $middlename,
+            ]);
+        }
+
+
+        $this->sendEmailRegistration($data['email'], $reg_key);
+
+        return json_encode([
+            'status'  => 'success',
+            'message' => $this->_('На указанную вами почту отправлены данные для входа в систему')
+        ]);
+    }
+
+
+    /**
      * @param $key
      * @param $password
      * @return false|string
      * @throws \Zend_Db_Adapter_Exception
+     * @throws \Zend_Db_Exception
      */
-    private function restorePass($key, $password) {
+    private function registrationComplete($key, $password) {
+
+        // Кастомное завершение регистрации
+        if ($this->core_config->registration->module) {
+            $module_name = strtolower($this->core_config->registration->module);
+            $location    = $this->getModuleLocation($module_name);
+
+            $mod_controller_name = "Mod" . ucfirst($module_name) . "Controller";
+
+            if ( ! file_exists("{$location}/{$mod_controller_name}.php")) {
+                throw new \Exception(sprintf($this->_('Контроллер модуля %s не найден'), $module_name));
+            }
+
+            require_once "{$location}/{$mod_controller_name}.php";
+
+            $this->setContext($module_name);
+            $mod_controller = new $mod_controller_name();
+            if ( ! ($mod_controller instanceof \Registration)) {
+                throw new \Exception(sprintf($this->_('Контроллер модуля %s не поддерживает регистрацию'), $module_name));
+            }
+
+            try {
+                $result_text = $mod_controller->coreRegistrationComplete($key, $password);
+
+                if ($result_text && is_string($result_text)) {
+                    return json_encode([
+                        'status'  => 'success',
+                        'message' => $result_text,
+                    ]);
+
+                } else {
+                    throw new \Exception($this->_('Не удалось завершить регистрацию. Попробуйте позже, либо свяжитесь с администратором'));
+                }
+
+            } catch (\Exception $e) {
+                return json_encode([
+                    'status'        => 'error',
+                    'error_message' => $e->getMessage(),
+                ]);
+            }
+        }
+
+
 
         $user_info = $this->db->fetchRow("
-            SELECT id,
-                   user_id
-            FROM mod_ordering_contractors 
+            SELECT u_id AS id,
+                   u_login AS login,
+                   email
+            FROM core_users 
             WHERE reg_key = ?
               AND date_expired > NOW()
             LIMIT 1
@@ -719,23 +797,118 @@ class Login extends Db {
 
         if (empty($user_info)) {
             return json_encode([
-                'status'  => 'error',
-                'message' => $this->_('Ссылка устарела')
+                'status'        => 'error',
+                'error_message' => $this->_('Ссылка устарела'),
             ]);
         }
 
-        $where = $this->db->quoteInto('id = ?', $user_info['id']);
-        $this->db->update('mod_ordering_contractors', [
+        $where = $this->db->quoteInto('u_id = ?', $user_info['id']);
+        $this->db->update('core_users', [
+            'visible'         => 'Y',
+            'is_pass_changed' => 'Y',
+            'u_pass'          => \Tool::pass_salt($password),
+            'reg_key'         => new \Zend_Db_Expr('NULL'),
+            'date_expired'    => new \Zend_Db_Expr('NULL'),
+        ], $where);
+
+
+        return json_encode([
+            'status'  => 'success',
+            'message' => '<h4>Готово!</h4><p>Вы можете войти в систему</p>'
+        ]);
+    }
+
+
+    /**
+     * @param $email
+     * @return false|string|string[]
+     * @throws \Zend_Db_Adapter_Exception
+     * @throws \Zend_Db_Exception
+     */
+    private function restore($email) {
+
+        $user_id = $this->db->fetchOne("
+            SELECT u.u_id
+            FROM core_users AS u
+            WHERE u.email = ?
+            LIMIT 1
+        ", $email);
+
+        if (empty($user_id)) {
+            return json_encode([
+                'status'        => 'error',
+                'error_message' => $this->_('Пользователя с таким Email нет в системе')
+            ]);
+        }
+
+
+        $reg_key = md5($email . microtime());
+        $where   = $this->db->quoteInto('u_id = ?', $user_id);
+        $this->db->update('core_users', [
+            'reg_key'      => $reg_key,
+            'date_expired' => new \Zend_Db_Expr('DATE_ADD(NOW(), INTERVAL 1 DAY)')
+        ], $where);
+
+
+        $name     = $this->config->system && $this->config->system->name ? $this->config->system->name : $_SERVER['SERVER_NAME'];
+        $protocol = ! empty($this->config->system) && $this->config->system->https ? 'https' : 'http';
+        $host     = ! empty($this->config->system) ? $this->config->system->host : '';
+        $doc_path = rtrim(DOC_PATH, '/') . '/';
+
+        $content_email = "
+            Вы запросили смену пароля на сервисе {$host}<br>
+            Для продолжения <b>перейдите по указанной ниже ссылке</b>.<br><br>
+
+            <a href=\"{$protocol}://{$host}{$doc_path}index.php?core=restore_complete&key={$reg_key}\" 
+               style=\"font-size: 16px\">{$protocol}://{$host}{$doc_path}index.php?core=restore_complete&key={$reg_key}</a>
+        ";
+
+        $this->setContext('queue');
+
+        require_once 'Email.php';
+        $core_email = new \Core2\Email();
+        $core_email->to($email)
+            ->subject("{$name}: Восстановление пароля")
+            ->body($content_email)
+            ->send(true);
+
+        return json_encode([
+            'status'  => 'success',
+            'message' => $this->_('На указанную вами почту отправлены данные для смены пароля')
+        ]);
+    }
+
+
+    /**
+     * @param $key
+     * @param $password
+     * @return false|string
+     * @throws \Zend_Db_Adapter_Exception
+     */
+    private function restoreComplete($key, $password) {
+
+        $user_id = $this->db->fetchOne("
+            SELECT u_id
+            FROM core_users 
+            WHERE reg_key = ?
+              AND date_expired > NOW()
+            LIMIT 1
+        ", $key);
+
+        if (empty($user_id)) {
+            return json_encode([
+                'status'        => 'error',
+                'error_message' => $this->_('Ссылка устарела')
+            ]);
+        }
+
+        $where = $this->db->quoteInto('u_id = ?', $user_id);
+        $this->db->update('core_users', [
+            'u_pass'       => \Tool::pass_salt($password),
             'reg_key'      => new \Zend_Db_Expr('NULL'),
             'date_expired' => new \Zend_Db_Expr('NULL'),
         ], $where);
 
-        if ( ! empty($user_info['user_id'])) {
-            $where = $this->db->quoteInto('u_id = ?', $user_info['user_id']);
-            $this->db->update('core_users', [
-                'u_pass' => \Tool::pass_salt($password),
-            ], $where);
-        }
 
         return json_encode([
             "status"  => "success",
@@ -751,8 +924,9 @@ class Login extends Db {
      */
     private function sendEmailRegistration($mail_address, $reg_key) {
 
-        $protocol = ! empty($this->config->system) && $this->config->system->https ? 'https' : 'http';
-        $host     = ! empty($this->config->system) ? $this->config->system->host : '';
+        $name     = $this->config->system && $this->config->system->name ? $this->config->system->name : $_SERVER['SERVER_NAME'];
+        $protocol = $this->config->system && $this->config->system->https ? 'https' : 'http';
+        $host     = $this->config->system ? $this->config->system->host : '';
         $doc_path = rtrim(DOC_PATH, '/') . '/';
 
         $content_email = "
@@ -768,7 +942,7 @@ class Login extends Db {
         require_once 'Email.php';
         $email = new \Core2\Email();
         $email->to($mail_address)
-            ->subject("Автопромсервис: Регистрация на сервисе")
+            ->subject("{$name}: Регистрация на сервисе")
             ->body($content_email)
             ->send(true);
     }
@@ -802,6 +976,133 @@ class Login extends Db {
 
 
     /**
+     * Получение данных дя пользователя root
+     * @return array
+     */
+    private final function getAuthRoot() {
+
+        require_once __DIR__ . '/../CoreController.php';
+
+        $auth            = [];
+        $auth['u_pass']  = \CoreController::RP;
+        $auth['u_id']    = -1;
+        $auth['u_login'] = 'root';
+        $auth['email']   = 'easter.by@gmail.com';
+
+        return $auth;
+    }
+
+
+    /**
+     * @param $login
+     * @param $password
+     * @return array
+     * @throws \Exception
+     */
+    private function getAuthLdap($login, $password) {
+
+        require_once 'LdapAuth.php';
+
+        $ldapAuth = new \LdapAuth();
+        $ldapAuth->auth($login, $password);
+        $ldapStatus = $ldapAuth->getStatus();
+
+        switch ($ldapStatus) {
+            case \LdapAuth::ST_LDAP_AUTH_SUCCESS :
+                $userData = $ldapAuth->getUserData();
+                $login    = $userData['login'];
+
+                if (isset($userData['root']) && $userData['root'] === true) {
+                    return $this->getAuthRoot();
+                }
+
+                $user_id = $this->dataUsers->fetchRow($this->db->quoteInto("u_login = ?", $login))->u_id;
+                if ( ! $user_id) {
+                    //create new user
+                    $this->db->insert('core_users', [
+                        'visible'     => 'Y',
+                        'is_admin_sw' => $userData['admin'] ? 'Y' : 'N',
+                        'u_login'     => $login,
+                        'role_id'     => $this->config->ldap->role_id ?: new \Zend_Db_Expr('NULL'),
+                        'date_added'  => new \Zend_Db_Expr('NOW()'),
+                    ]);
+
+                } elseif ($userData['admin']) {
+                    $this->db->update('core_users', ['is_admin_sw' => 'Y'], $this->db->quoteInto('u_id = ?', $user_id));
+                }
+
+                return $this->getAuthLogin($login);
+                break;
+
+            case \LdapAuth::ST_LDAP_USER_NOT_FOUND :
+                //удаляем пользователя если его нету в AD и с префиксом LDAP_%
+                //$this->db->query("DELETE FROM core_users WHERE u_login = ?", $login);
+                break;
+
+            case \LdapAuth::ST_LDAP_INVALID_PASSWORD :
+                throw new \Exception($this->translate->tr("Неверный пароль или пользователь отключён"));
+                break;
+
+            case \LdapAuth::ST_ERROR :
+                throw new \Exception($this->translate->tr("Ошибка LDAP: ") . $ldapAuth->getMessage());
+                break;
+
+            default:
+                throw new \Exception($this->translate->tr("Неизвестная ошибка авторизации по LDAP"));
+                break;
+        }
+
+        return [];
+    }
+
+
+    /**
+     * @param $login
+     * @return mixed
+     */
+    private function getAuthLogin($login) {
+
+        return $this->dataUsers->getUserByLogin($login);
+    }
+
+
+    /**
+     * Сохранение информации о входе пользователя
+     * @param SessionContainer $auth
+     * @return mixed
+     * @throws \Exception
+     */
+    private function storeSession(SessionContainer $auth) {
+
+        if ($auth && $auth->ID && $auth->ID > 0) {
+
+            $sid = $auth->getManager()->getId();
+            $sess = $this->dataSession;
+            $row = $sess->fetchRow($sess->select()
+                ->where("logout_time IS NULL AND user_id = ?", $auth->ID)
+                ->where("sid = ?", $sid)
+                ->where("ip = ?", $_SERVER['REMOTE_ADDR'])
+                ->limit(1));
+
+            if ( ! $row) {
+                $row             = $sess->createRow();
+                $row->sid        = $sid;
+                $row->login_time = new \Zend_Db_Expr('NOW()');
+                $row->user_id    = $auth->ID;
+                $row->ip         = $_SERVER['REMOTE_ADDR'];
+                $row->save();
+            }
+
+            if ( ! $row->id) {
+                throw new \Exception($this->translate->tr("Не удалось сохранить данные сессии"));
+            }
+
+            return $row->id;
+        }
+    }
+
+
+    /**
      * @return string
      * @throws \Exception
      */
@@ -819,26 +1120,6 @@ class Login extends Db {
 
         $tpl->assign('favicon.png', isset($this->favicon['png']) && is_file($this->favicon['png']) ? $this->favicon['png'] : '');
         $tpl->assign('favicon.ico', isset($this->favicon['ico']) && is_file($this->favicon['ico']) ? $this->favicon['ico'] : '');
-
-
-        $system_js = "";
-        if ($this->config->mail &&
-            $this->config->mail->server &&
-            (($this->core_config->registration && $this->core_config->registration->on) ||
-            ($this->core_config->restore && $this->core_config->restore->on))
-        ) {
-            $scripts = [
-                'core2/js/core-login.js',
-                'core2/js/jquery.maskedinput.min.js',
-            ];
-            foreach ($scripts as $src) {
-                if (file_exists($src)) {
-                    $system_js .= "<script src=\"{$src}\"></script>";
-                }
-            }
-        }
-        $tpl->assign("<!--system_js-->", $system_js);
-
 
         return $tpl->render();
     }
