@@ -4,7 +4,7 @@ namespace Core2;
 require_once DOC_ROOT . "/core2/inc/classes/Common.php";
 require_once DOC_ROOT . "/core2/inc/classes/class.list.php";
 
-use Zend\Session\Container as SessionContainer;
+use Laminas\Session\Container as SessionContainer;
 use GuzzleHttp\Client;
 use GuzzleHttp\Psr7;
 use GuzzleHttp\Exception\RequestException;
@@ -517,6 +517,8 @@ class InstallModule extends \Common {
         $this->installSql();
         //копирование файлов
         $this->copyModFiles();
+        //удовлетворяем зависимости модуля
+        $this->resolveDependencies(true);
         //инфа о модуле
         $arrForInsert['module_id']       = $this->mInfo['install']['module_id'];
         $arrForInsert['m_name']          = $this->mInfo['install']['module_name'];
@@ -721,6 +723,8 @@ class InstallModule extends \Common {
         $this->migrateSql();
         //копируем файлы из архива
         $this->copyModFiles();
+        //удовлетворяем зависимости модуля
+        $this->resolveDependencies();
         //инфа о модуле
         //$arrForUpgrade['m_name']        = $this->mInfo['install']['module_name'];
         $arrForUpgrade['lastuser']        = $this->lastUser;
@@ -1397,10 +1401,11 @@ class InstallModule extends \Common {
                 return json_decode($body);
             }
         } catch (RequestException $e) {
-            $msg = Psr7\str($e->getRequest());
+            $msg = '';
             if ($e->hasResponse()) {
                 $msg = Psr7\str($e->getResponse());
             }
+            if (!$msg) $msg = $e->getMessage();
             throw new \Exception($msg);
         }
 
@@ -1415,7 +1420,7 @@ class InstallModule extends \Common {
      * @return  mixed                   Массив с информацией о доступных для установки модулей
      * @throws  \Exception
      */
-    public function getModsListFromRepo($repo_url) {
+    private function getModsListFromRepo($repo_url) {
         //проверяем есть ли ключ к репозиторию, если нет то получаем
         $repo_url = trim($repo_url);
         if (substr_count($repo_url, "repo?apikey=") == 0) {
@@ -1493,14 +1498,17 @@ class InstallModule extends \Common {
     /**
      * Таблица-список доступных модулей из репозитория
      *
-     * @param   string  $repo_url   Подготовленный URL для запроса к репозиторию
+     * @param   string  $repo_id   порядковый номер репозитория
      *
      */
-    public function getHTMLModsListFromRepo($repo_url) {
-        $_GET['repo_id'] = !empty($_GET['repo_id']) ? $_GET['repo_id'] : "";
+    public function getHTMLModsListFromRepo($repo_id) {
         try {
+            $mod_repos = $this->getSetting('repo');
+            $mod_repos = explode(";", $mod_repos);
+            $repo_url = $mod_repos[$repo_id];
             //достаём список модулей
             $repo_list = $this->getModsListFromRepo($repo_url);
+            if (!$repo_list) throw new \RuntimeException(404);
 
             $api_key = explode("?apikey=", $repo_url);
             $api_key = !empty($api_key[1]) ? $api_key[1] : uniqid();
@@ -1653,8 +1661,8 @@ class InstallModule extends \Common {
                 $copy_list[$module_id] = $val[$max_ver];
                 unset($val[$max_ver]);
                 if (!empty($val)) {
-                    $copy_list[$module_id]['version'] .= " <a href=\"\" onclick=\"$('.repo_table_{$_GET['repo_id']}_{$module_id}').toggle(); return false;\">Предыдущие версии</a><br>";
-                    $copy_list[$module_id]['version'] .= "<table width=\"100%\" class=\"repo_table_{$_GET['repo_id']}_{$module_id}\" style=\"display: none;\"><tbody>";
+                    $copy_list[$module_id]['version'] .= " <a href=\"\" onclick=\"$('.repo_table_{$repo_id}_{$module_id}').toggle(); return false;\">Предыдущие версии</a><br>";
+                    $copy_list[$module_id]['version'] .= "<table width=\"100%\" class=\"repo_table_{$repo_id}_{$module_id}\" style=\"display: none;\"><tbody>";
                     foreach ($val as $version => $val2) {
                         $copy_list[$module_id]['version'] .= "<tr><td style=\"border: 0px; padding: 0px;\">{$version}</td><td style=\"border: 0px; text-align: right; padding: 0px;\">{$val2['install_info']}</td></tr>";
                     }
@@ -1667,7 +1675,7 @@ class InstallModule extends \Common {
 //            $list->setRecordCount($per_page);
 
             //пагинация
-            $ss         = new \Zend_Session_Namespace('Search');
+            $ss         = new SessionContainer('Search');
             $ssi        = 'main_' . $list_id;
             $ss         = $ss->$ssi;
             $per_page   = empty($ss["count_{$list_id}"]) ? 1 : (int)$ss["count_{$list_id}"];
@@ -1689,7 +1697,12 @@ class InstallModule extends \Common {
             $list->data = $copy_list;
             $list->showTable();
 
-        } catch (\Exception $e) {
+        }
+        catch (\RuntimeException $e) {
+            $this->addNotice("", $this->translate->tr("При подключении к репозиторию произошла ошибка"), $e->getMessage(), "danger");
+            echo $this->printNotices();
+        }
+        catch (\Exception $e) {
             $this->addNotice("", $this->translate->tr("При подключении к репозиторию произошла ошибка"), $e->getMessage(), "danger");
             echo $this->printNotices();
         }
@@ -2527,6 +2540,34 @@ class InstallModule extends \Common {
             }
         }
         return $files;
+    }
+
+    /**
+     * Установка зависимостей composer
+     * @param bool $install
+     * @throws \Exception
+     */
+    private function resolveDependencies($install = false) {
+        //return;
+        //is composer.json exists
+        if (file_exists($this->installPath . DIRECTORY_SEPARATOR . "composer.json")) {
+            chdir($this->installPath);
+            $output = "";
+            $return_var = "";
+            $msg = "";
+            exec("composer update 2>&1", $output, $return_var);
+            foreach ($output as $k => $item) {
+                $item = trim($item);
+                if ($item == '[RuntimeException]') {
+                    //$msg = $output[$k + 1];
+                    $notice = $this->translate->tr("Не удалось провести установку. Попробуйте установить зависимости вручную.");
+                    $this->addNotice($this->translate->tr("Файлы модуля"), $this->translate->tr("Проверка зависимостей: "), $notice, "danger");
+                }
+            }
+            if ($msg) {
+                throw new \Exception($msg);
+            }
+        }
     }
 
 }
