@@ -1,6 +1,6 @@
 <?php
 
-use Zend\Session\Container as SessionContainer;
+use Laminas\Session\Container as SessionContainer;
 
 
 require_once("core2/inc/ajax.func.php");
@@ -472,7 +472,8 @@ class ModAjax extends ajaxFunc {
 
         $core_config            = \Zend_Registry::getInstance()->get('core_config');
         $is_auth_certificate_on = $core_config->auth && $core_config->auth->x509 && $core_config->auth->x509->on;
-        $is_auth_pass_on        = $core_config->auth && $core_config->auth->pass && $core_config->auth->pass->on;
+        $is_auth_pass_on        = true;
+        if ($core_config->auth) $is_auth_pass_on        = $core_config->auth && $core_config->auth->pass && $core_config->auth->pass->on;
         $is_auth_ldap_on        = $this->config->ldap && $this->config->ldap->active;
 
         $refid  = $this->getSessFormField($data['class_id'], 'refid');
